+++ common.py
@@ -1,7 +1,9 @@
 """
 Common classes to wrap around pdsh (parallel shell)
 """
+import copy
 import errno
+import itertools
 import logging
 import os
 import signal
@@ -11,6 +13,32 @@ import subprocess
 import settings
 
 logger = logging.getLogger("cbt")
+
+def all_configs(config):
+    """
+    return all parameter combinations for config
+    config: dict - list of params
+    iterate over all top-level lists in config
+    """
+    cycle_over_lists = []
+    cycle_over_names = []
+    default = {}
+
+    for param, value in list(config.items()):
+        # acceptable applies to benchmark as a whole, no need to it to
+        # the set for permutation
+        if param == "acceptable":
+            default[param] = value
+        elif isinstance(value, list):
+            cycle_over_lists.append(value)
+            cycle_over_names.append(param)
+        else:
+            default[param] = value
+
+    for permutation in itertools.product(*cycle_over_lists):
+        current = copy.deepcopy(default)
+        current.update(list(zip(cycle_over_names, permutation)))
+        yield current
 
 class Localhost(object):
     """
